requirements.txt
snowflake-connector-python==2.0.4